pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.18",
    "anchorpy>=0.21.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "qrcode[pil]>=8.2",
    "requests>=2.32.3",
    "solana>=0.36.6",
    "telegram>=0.0.1",
]
